feat(tiers): add option to inherit models when creating a new tier
This commit introduces the ability to inherit models from the previous tier when creating a new tier. The `useTiers` hook and `TierCreationSheet` component have been updated to support this feature. Additionally, the `@radix-ui/react-checkbox` dependency has been updated to version 1.2.3, and minor style adjustments were made to the checkbox component.